pyproject.toml
[project]
name = "realtimemap"
version = "0.1.0"
description = "RealTime-map backend"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "fastapi-babel>=1.0.0",
    "fastapi-cache2[redis]>=0.2.2",
    "fastapi-pagination>=0.13.1",
    "fastapi-users[oauth,sqlalchemy]>=14.0.1",
    "fasteners>=0.19",
    "geoalchemy2>=0.17.1",
    "geojson-pydantic>=2.0.0",
    "orjson>=3.10.18",
    "phonenumbers>=9.0.7",
    "pillow>=11.2.1",
    "pydantic-extra-types>=2.10.5",
    "pydantic-settings>=2.9.1",
    "shapely>=2.1.1",
    "sqlalchemy>=2.0.41",
    "sqlalchemy-file>=0.6.0",
    "starlette-admin>=0.15.1",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
    "python-socketio[asyncio-client]>=5.13.0",
    "gunicorn>=23.0.0",
    "celery>=5.5.3",
    "asgiref>=3.9.0",
    "psycopg2-binary>=2.9.10",
    "fastapi-limiter>=0.1.6",
    "flower>=2.0.1",
    "pygeohash>=3.2.0",
    "itsdangerous>=2.2.0",
    "black>=25.1.0",
    "yookassa>=3.7.0",
    "redis>=4.6.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "icecream>=2.1.4",
    "pre-commit>=4.2.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-multipart>=0.0.20",
    "ruff>=0.11.10",
    "trio>=0.30.0",
]
win = [
    "psycopg2>=2.9.11",
]

[tool.mypy]
strict = true
